+++ src/lisp/kernel/cmp/opt/opt-array.lisp
@@ -95,3 +95,8 @@
             ;; unknown uaet, give up
             form))
       form))
+
+(define-compiler-macro bit (bit-array &rest indices)
+  `(aref (core::the-single (array bit) ,bit-array) ,@indices))
+(define-compiler-macro sbit (bit-array &rest indices)
+  `(aref (core::the-single (simple-array bit) ,bit-array) ,@indices))